remove old reocrder proxy and rename file
