+++ packages/recorder/controllers/proxy_controller.ts
@@ -1,92 +1,137 @@
 import express from 'express';
-import {SessionManager} from '../session_recording.ts/sessions_manager';
-import {RecordedCall} from '../src/types';
+import proxy from 'express-http-proxy';
+import {IncomingHttpHeaders} from 'http';
 
-import fetch from 'node-fetch';
+import {AppDependencies, RecordedCall} from '../src/types';
 
-type Deps = {sessionManager: SessionManager};
+const HTTP_HEADER_ORIGINAL_HOST = 'ORIGINAL_HOST';
 
-export const initProxyRouter = (deps: Deps) => {
+const mockedResponseDataCache: Record<string, object> = {
+    '/api/repos': {
+        items: [
+            {name: 'Some name'},
+        ],
+    },
+};
+
+const getCachedData = (req: express.Request, deps: AppDependencies): any => {
+    return mockedResponseDataCache[req.originalUrl] || '';
+};
+
+type ResponseInfo = {
+    headers: IncomingHttpHeaders;
+    statusCode?: number;
+}
+
+const recordRequest = (proxyRes: ResponseInfo, proxyResBody: string, req: express.Request, deps: AppDependencies) => {
+    const {original_host, ...reqHeaders} = req.headers;
+
+    const logErr = (message: string) => {
+        console.error(`Error recording request: ${message}`);
+    }
+
+    if (!original_host) {
+        logErr('No original_host header provided');
+        return;
+    }
+
+    if (typeof original_host !== 'string') {
+        logErr('Provided original_host header is not a string');
+        return;
+    }
+
+    let host = original_host;
+    if (!host.startsWith('https://') && !host.startsWith('http://')) {
+        host = 'https://' + host;
+    }
+
+    if (host.endsWith('/')) {
+        host = host.slice(0, host.length - 1);
+    }
+
+    const reqPath = req.originalUrl;
+    const destinationUrl = host + reqPath;
+
+    console.log(`FULLCIRCLE LOG: Proxying to url: ${destinationUrl}`);
+
+    // const fetchRes = await fetch(destinationUrl, {
+    //     headers,
+    // });
+
+    let responseBody = proxyResBody;
+    const responseHeaders = proxyRes.headers;
+
+    // const responseHeaders = fetchRes.headers;
+    // let responseBody: string | object = await fetchRes.text();
+    // let isJSON = false;
+    try {
+        responseBody = JSON.parse(responseBody);
+        // isJSON = true;
+    } catch (e) {
+        console.log('Tried to JSON parse response, but failed. Assuming response is not JSON.');
+    }
+
+    const call: RecordedCall = {
+        time: new Date().toISOString(),
+        host,
+        requestMethod: req.method,
+        requestPath: reqPath,
+        requestBody: req.body,
+        responseBody,
+        requestHeaders: req.headers,
+        responseHeaders,
+        requestIp: req.ip || '',
+        status: proxyRes.statusCode || 0,
+    }
+
+    deps.sessionManager.getCurrentSession()?.addCallToSession(call);
+
+    // if (isJSON) {
+    //     res.json(responseBody);
+    // } else {
+    //     res.send(responseBody);
+    // }
+}
+
+export const initProxyRouter = (deps: AppDependencies) => {
     const proxyRouter = express.Router();
 
-    proxyRouter.use('*', async (req, res) => {
+    proxyRouter.use((req, res, next) => {
         const {original_host, ...reqHeaders} = req.headers;
 
-        if (!original_host) {
-            res.json({error: 'No original_host header provided'});
+        if (!deps.shouldProxy) {
+            const data = getCachedData(req, deps);
+
+            recordRequest({headers: {}, statusCode: 200}, JSON.stringify(data), req, deps);
+
+            res.json(data);
             return;
         }
 
-        if (typeof original_host !== 'string') {
-            res.json({error: 'Provided original_host header is not a string'});
+        if (!original_host || typeof original_host !== 'string') {
+            res.json({error: `Please provide a host to proxy to with the HTTP header ${HTTP_HEADER_ORIGINAL_HOST}`});
             return;
         }
 
-        let host = original_host;
-        if (!host.startsWith('https://') && !host.startsWith('http://')) {
-            host = 'https://' + host;
-        }
+        proxy(original_host, {
+            proxyReqPathResolver: async (req) => {
+                // Optionally rewrite the path going out
+                return req.url;
+            },
 
-        if (host.endsWith('/')) {
-            host = host.slice(0, host.length - 1);
-        }
+            // proxyReqBodyDecorator: function (bodyContent, srcReq) {
+            //     return srcReq.rawBody || bodyContent;
+            // },
 
-        const headers: Record<string, string> = {};
-        for (const headerName of Object.keys(reqHeaders)) {
-            const headerValue = reqHeaders[headerName];
-            if (!headerValue) {
-                continue;
-            }
+            // eslint-disable-next-line @typescript-eslint/no-unused-vars
+            userResDecorator: (proxyRes, proxyResData, userReq, userRes) => {
+                const resBody = proxyResData.toString('utf8') as string;
+                recordRequest(proxyRes, resBody, userReq, deps);
 
-            if (typeof headerValue === 'string') {
-                headers[headerName] = headerValue;
-            } else {
-                headers[headerName] = headerValue.join(',');
-            }
-        }
+                return proxyResData;
+            },
 
-        const reqPath = req.originalUrl;
-        const destinationUrl = host + reqPath;
-
-        console.log(`FULLCIRCLE LOG: Proxying to url: ${destinationUrl}`);
-
-        const fetchRes = await fetch(destinationUrl, {
-            headers,
-        });
-
-        const responseHeaders = fetchRes.headers;
-        let responseBody: string | object = await fetchRes.text();
-        let isJSON = false;
-        try {
-            responseBody = JSON.parse(responseBody);
-            isJSON = true;
-        } catch (e) {
-            console.log('Tried to JSON parse response, but failed. Assuming response is not JSON.');
-        }
-
-        const call: RecordedCall = {
-            time: new Date().toISOString(),
-            host,
-            requestMethod: req.method,
-            requestPath: reqPath,
-            requestBody: req.body,
-            responseBody,
-            requestHeaders: headers,
-            responseHeaders: Array.from(responseHeaders).reduce<Record<string, string>>((accum, current) => {
-                accum[current[0]] = current[1];
-                return accum;
-            }, {}),
-            requestIp: req.ip || '',
-            status: fetchRes.status,
-        }
-
-        deps.sessionManager.getCurrentSession()?.addCallToSession(call);
-
-        if (isJSON) {
-            res.json(responseBody);
-        } else {
-            res.send(responseBody);
-        }
+        })(req, res, next);
     });
 
     return proxyRouter;
